backend/requirements.txt
Django>=5.0
djangorestframework
neomodel
django-neomodel
gunicorn
neo4j